Move the `value_utils` module into the `federated_context` package.
PiperOrigin-RevId: 358021918
